Theme Rework
Added skeleton-auto.css, which dynamically sets the theme based on the browser and/or the Operating Systems colors, if supported.

Set Default Theme to Auto with comments on helping switch to light and dark theme stylesheets.

Darkened the skeleton-dark.css background color to the same as the color that skeleton-auto.css sets it as, which is html recommended standard color of #121212.